requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
huggingface-hub==0.13.3
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
onnxruntime==1.14.1
packaging==23.0
protobuf==4.22.1
pyreadline3==3.4.1
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
